+++ Maple2.Server.Game/Session/GameSession.cs
@@ -185,7 +185,7 @@ public sealed partial class GameSession : Core.Network.Session {
         Config.LoadKeyTable();
         // GuideRecord
         // DailyWonder*
-        Send(GameEventPacket.Load(db.GetEvents().ToList()));
+        Send(GameEventPacket.Load(db.GetEvents()));
         // BannerList
         // RoomDungeon
         // FieldEntrance
